[#3119] static cast time_t to uint32

+++ src/lib/dhcpsrv/csv_lease_file6.cc
@@ -221,7 +221,7 @@ CSVLeaseFile6::readCltt(const CSVRow& row) {
     time_t cltt =
         static_cast<time_t>(row.readAndConvertAt<uint64_t>(getColumnIndex("expire"))
                             - readValid(row));
-    return (cltt);
+    return (static_cast<uint32_t>(cltt));
 }
 
 SubnetID
